Create CurrencyConveter.java
Service Code (Business Logic):

The web service implemented performs a range of simple conversions between different currencies. The application converts a value from one currency to another; for example, it can be used to convert ten pounds sterling into an equivalent sum in American dollars or any one of a range of other currencies.

This is the Java code for the web service business logic, namely CurrencyConveter.java, attached as a second file of the repository.